Fixed stats not being calculated for logfile
Previously, complexity was set just if `use_arch=true` or one of
`sel` or `surv` parameters would use a pareto dominance criteria
(because they are calculated only if strictly needed).

If we want a logfile, but `use_arch=false` and no pareto dominance
is calculated in `sel` or `surv`, then the logfile would have
all zeros for complexity statistics.

Now we make sure we calculate the stats if they are being reported
to the user, regardless if is in archive, in verbosity, or in
a logfile.