Fixes stats being reported before updating the objectives
Previously, I made a change to calculate the complexity before logging,
as the complexity column was filled with zeros if we dont use the
NSGA2 survival.

Still, I was updating the stats after sending them to the log file.

Now Im doing everything in the right order, and the log file doesnt
get filled with zeros in complexity.