Sklearn changed check_estimators_dtypes and now fails the test
FEAT is designed to work with continuous class encodings, and do
not perform the transformation itself.

Instead of changing the behavior of FEAT, I think the error message
is already clear enough and the user can easily get around this
problem if the data is not as FEAT expects.

For the record, this is the new test function:

```python
def check_estimators_dtypes(name, estimator_orig):
    rnd = np.random.RandomState(0)
    X_train_32 = 3 * rnd.uniform(size=(20, 5)).astype(np.float32)
    X_train_32 = _enforce_estimator_tags_X(estimator_orig, X_train_32)
    X_train_64 = X_train_32.astype(np.float64)
    X_train_int_64 = X_train_32.astype(np.int64)
    X_train_int_32 = X_train_32.astype(np.int32)
    y = np.array([1, 2] * 10, dtype=np.int64)
    y = _enforce_estimator_tags_y(estimator_orig, y)
```

and FEAT wont work due to:
```python
self.classes_ = [int(i) for i in np.unique(np.asarray(y))]
        if (any([
            i != j
            for i,j in zip(self.classes_, np.arange(np.max(self.classes_)))
        ])):
            raise ValueError('y must be a contiguous set of labels from ',
                             '0 to n_classes. y contains the values {}'.format(
                                self.classes_)
                            )
```.